refactor: split `utils.mjs`, move `getUsernameFromArgv` to `arguments.mjs`
